pyproject.toml
[project]
name = "langchain-basics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.7.4",
    "fastapi>=0.110.0",
    "google-generativeai>=0.5.0",
    "groq>=0.5.0",
    "ipykernel>=6.0.0",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-google-genai>=1.0.0",
    "langchain-groq>=0.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.55.0",
    "uvicorn>=0.29.0",
]
